Added feature: add feed by url
