+++ src/models/podcast_rssadd_model.rs
@@ -1,0 +1,5 @@
+#[derive(Debug, Serialize, Deserialize,Clone)]
+pub struct PodcastRSSAddModel{
+    #[serde(rename = "rssFeedUrl")]
+    pub rss_feed_url: String,
+}
