+++ ui/src/language/json/en.json
@@ -104,5 +104,7 @@
   "os": "OS",
   "podcast-refreshed": "Podcast {{name}} refreshed.",
   "version-infos": "Version Info",
-  "progress": "Progress"
+  "progress": "Progress",
+  "opml-file": "OPML file",
+  "rss-feed-url": "RSS feed URL"
 }
