+++ ui/src/models/AddTypes.ts
@@ -1,0 +1,6 @@
+export enum AddTypes {
+    ITUNES = "itunes",
+    PODINDEX = "podindex",
+    OPML = "opml",
+    FEED = "feed"
+}
